Move config exception to dedicated classes with proper namespace
